requirements.txt
django==3.2.3
pillow==8.2.0
django-import-export==2.5.0